+++ zorphy/lib/src/generators/generators.dart
@@ -2,11 +2,12 @@
 export 'base_generator.dart';
 export 'class_declaration_generator.dart';
 export 'copywith_generator.dart';
-export 'factory_method_generator.dart';
 export 'equals_tostring_generator.dart';
-export 'patch_generator.dart';
-export 'json_generator.dart';
 export 'extension_generator.dart';
+export 'factory_method_generator.dart';
 export 'fields_class_generator.dart';
+export 'json_generator.dart';
+export 'patch_generator.dart';
 export 'property_helper_generator.dart';
 export 'agent_directive_generator.dart';
+export 'value_equality_generator.dart';